+++ lib/public/files/objectstore/iobjectstore.php
@@ -5,6 +5,11 @@ namespace OCP\Files\ObjectStore;
 interface IObjectStore {
 
 	/**
+	 * @return string the container or bucket name where objects are stored
+	 */
+	function getStorageId();
+
+	/**
 	 * @param string $urn the unified resource name used to identify the object
 	 * @param string $tmpFile path to the local temporary file that should be
 	 *        used to store the object
@@ -12,6 +17,7 @@ interface IObjectStore {
 	 * @throws Exception when something goes wrong, message will be logged
 	 */
 	function getObject($urn, $tmpFile);
+
 	/**
 	 * @param string $urn the unified resource name used to identify the object
 	 * @param string $tmpFile path to the local temporary file that the object
@@ -19,8 +25,7 @@ interface IObjectStore {
 	 * @return void
 	 * @throws Exception when something goes wrong, message will be logged
 	 */
-	function updateObject($urn, $tmpFile = null);
-
+	function writeObject($urn, $tmpFile = null);
 
 	/**
 	 * @param string $urn the unified resource name used to identify the object
